requirements.txt
# The base set of requirements for the registry less than these. However, for 
# the purposes of development and testing, I've included additional packages 
# that are useful to install.

behave==1.2.5
behave-django==1.1.0
certifi==2018.1.18
chardet==3.0.4
Django==2.0.1
django-debug-toolbar==1.9.1
django-extensions==1.9.9
djangorestframework==3.7.7
flake8==3.5.0
idna==2.6
Markdown==2.6.11
mccabe==0.6.1
parse==1.8.2
parse-type==0.4.2
pep8==1.7.1
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2017.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
typing==3.6.2
urllib3==1.22
Werkzeug==0.14.1
yapf==0.20.1
